--- src/js/get-news.js
@@ -1,27 +1,0 @@
-import {
-    clearMainContainer,
-    addElementToMainContainer,
-} from './create-container.js';
-
-export async function getNews(source = 'google-news') {
-    try {
-        const response = await fetch(
-            `https://newsapi.org/v1/articles?source=${source}&apiKey=8d9cd797251f4fc4a24f983a029f67cc`
-        );
-        if (response.status !== 200) {
-            throw new Error(
-                `Error : ${response.statusText}, status: ${response.status}`
-            );
-        }
-        const data = await response.json();
-        if (data.status === 'error' && !data['articles']) {
-            throw new Error(`Error : ${data.message}`);
-        }
-        clearMainContainer();
-        addElementToMainContainer('news', data['articles']);
-    } catch (err) {
-        clearMainContainer();
-        addElementToMainContainer('error');
-        console.log(err.message);
-    }
-}
