--- src/js/toggle.js
@@ -1,4 +1,0 @@
-export function navToggle(element) {
-    document.querySelector('.nav-item.active').classList.remove('active');
-    element.classList.add('active');
-}
